+++ tests/default-ui.test.tsx
@@ -81,6 +81,9 @@ const processSummary: ProcessSummary = {
   total_slots: 4,
   assigned_slots: 1,
   available_slots: 3,
+  balanced_participant_count: 0,
+  pending_participant_count: 0,
+  overloaded_participant_count: 1,
   current_turn: {
     meeting_session_id: "22222222-2222-4222-8222-222222222222",
     selection_turn_id: "33333333-3333-4333-8333-333333333333",